starting simulator update
