+++ setup.cfg
@@ -46,15 +46,15 @@ keywords =
 
 [options]
 zip_safe = True
-python_requires = >=3.9
+python_requires = >=3.11
 install_requires =
     sbmlutils>=0.8.7
-    numpy==1.21.3         # see https://github.com/sys-bio/roadrunner/issues/963
-    libroadrunner==2.1.3  # see https://github.com/sys-bio/roadrunner/issues/963
-    scipy==1.10.1
-    pint==0.19.2
+    numpy>=1.26.4
+    libroadrunner>=2.6.0
+    scipy>=1.12.0
+    pint>=0.23
 
-    # -> sbmlutils (no version pin)
+    # from sbmlutils (no version pin)
     python-libsbml
     rich
     depinfo
@@ -66,28 +66,22 @@ install_requires =
     matplotlib
 
     # dependencies
-    petab>=0.1.25
-    # fix for windows access violations
-	ray[default]>=2.9.3
-    grpcio==1.49.1
-    pydantic>=2.6.3
-	tables>=3.8.0
-	xarray>=2023.6.0
-	bottleneck>=1.3.4
+    petab>=0.2.9
+    pydantic>=2.6.4
+	tables>=3.9.2
+	xarray>=2024.3.0
+	bottleneck>=1.3.8
+	psutil>=5.9.8
+	setproctitle>=1.3.3
+
+	plotly>=5.20.0
+	altair>=5.3.0
+	seaborn>=0.13.2
+    xmltodict>=0.13.0
+	pyDOE>=0.3.8
 
 	python-libsedml>=2.0.32
 	python-libnuml>=1.1.6
-    # biosimulators-utils>=0.1.171
-
-	psutil>=5.9.0
-	setproctitle>=1.2.2
-
-	plotly>=5.6.0
-	altair>=4.2.0
-	seaborn>=0.11.2
-
-    xmltodict>=0.12.0
-	pyDOE>=0.3.8
 tests_require =
 	tox>=3.24.3
 	pytest>=7.0.1
