+++ src/sbmlsim/examples/experiments/repressilator/repressilator_scans.py
@@ -17,7 +17,7 @@ from sbmlsim.simulation import (
     Timecourse,
     TimecourseSim,
 )
-from sbmlsim.simulator.simulation_ray import SimulatorParallel, SimulatorSerial
+from sbmlsim.simulator.simulation_serial import SimulatorSerial
 from sbmlsim.task import Task
 from sbmlsim.test import MODEL_REPRESSILATOR
 
@@ -219,7 +219,7 @@ def run_repressilator_experiments(output_path: Path) -> Path:
     base_path = Path(__file__).parent
     data_path = base_path
 
-    for simulator in [SimulatorSerial(), SimulatorParallel()]:
+    for simulator in [SimulatorSerial()]:
         runner = ExperimentRunner(
             [RepressilatorScanExperiment],
             simulator=simulator,
