+++ src/sbmlsim/simulator/__init__.py
@@ -1,3 +1,2 @@
 """Package for simulator."""
-from .simulation import SimulatorWorker
 from .simulation_serial import SimulatorSerial
