+++ src/sbmlsim/test/test_simulator.py
@@ -3,7 +3,6 @@ import roadrunner
 
 from sbmlsim.model import RoadrunnerSBMLModel
 from sbmlsim.simulator import SimulatorSerial
-from sbmlsim.simulator.simulation_ray import SimulatorParallel
 from sbmlsim.test import MODEL_REPRESSILATOR
 
 
@@ -36,30 +35,4 @@ def test_tolerances_serial2():
         relative_tolerance=rel_tol,
     )
     assert isinstance(simulator, SimulatorSerial)
-    _tolerance_test(r=simulator.model.r, abs_tol=abs_tol, rel_tol=rel_tol)
-
-
-def test_tolerances_parallel():
-    abs_tol = 1e-14
-    rel_tol = 1e-14
-
-    simulator = SimulatorParallel(
-        model=MODEL_REPRESSILATOR,
-        absolute_tolerance=abs_tol,
-        relative_tolerance=rel_tol,
-    )
-    assert isinstance(simulator, SimulatorParallel)
-    _tolerance_test(r=simulator.model.r, abs_tol=abs_tol, rel_tol=rel_tol)
-
-
-def test_tolerances_parallel2():
-    abs_tol = 1e-14
-    rel_tol = 1e-14
-
-    simulator = SimulatorParallel(
-        model=RoadrunnerSBMLModel(source=MODEL_REPRESSILATOR),
-        absolute_tolerance=abs_tol,
-        relative_tolerance=rel_tol,
-    )
-    assert isinstance(simulator, SimulatorParallel)
     _tolerance_test(r=simulator.model.r, abs_tol=abs_tol, rel_tol=rel_tol)
